refactor(web): isolate Lambda ast-grep relay
